+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 2.1.0
+
+- Added SetSaslCredentials. This new method (on the Producer, Consumer, and AdminClient) allows modifying the stored
+  SASL PLAIN/SCRAM credentials that will be used for subsequent (new) connections to a broker (#1980).
+
+
 # 2.0.2
 
 ## Upgrade considerations